Refactor sandbox execution code
