+++ src/openelm/sandbox/server/index.py
@@ -1,10 +1,8 @@
-import json
-
-from flask import Flask, abort, jsonify, request
+from flask import Flask, request
 from numpy import ndarray
 
 from .environments.walker.walk_creator import Walker
-from .utils import sandbox_unsafe_execute
+from .sandbox_codex_execute import unsafe_execute
 
 app = Flask(__name__)
 
@@ -15,10 +13,10 @@ def bad_request(message, **kwargs):
 
 def generate_racer(code_str, timeout):
     try:
-        execution_result = sandbox_unsafe_execute(code_str, "make_walker", timeout)
-    except:
+        execution_result = unsafe_execute(code_str, "make_walker", timeout)
+    except Exception:
         return bad_request(
-            f"failed to execute code", unsafe_execute_error_code=6
+            "Failed to execute code", unsafe_execute_error_code=6
         )  # 6: Other errors.
     if isinstance(execution_result, Walker):
         sodaracer_dict: dict = execution_result.to_dict()
@@ -30,36 +28,38 @@ def generate_racer(code_str, timeout):
         else:
             # 1: Code runs but fails a test (valid walker).
             return bad_request(
-                f"invalid walker",
+                "Invalid walker",
                 walker=execution_result.to_dict(),
                 unsafe_execute_error_code=1,
             )
     elif isinstance(execution_result, int):
         return bad_request(
-            f"failed sandbox_unsafe_execute", unsafe_execute_error_code=execution_result
+            "Failed sandbox_unsafe_execute", unsafe_execute_error_code=execution_result
         )
     else:
         return bad_request(
-            f"not walker", unsafe_execute_error_code=1
+            "Not walker", unsafe_execute_error_code=1
         )  # 1: Code runs but fails a test (valid walker).
 
 
 @app.route("/gen_racer", methods=["POST"])
 def gen_racer():
     req_json = request.get_json()
-    if isinstance(req_json["code"], list):
-        return [
-            generate_racer(code, req_json["timeout"])[0] for code in req_json["code"]
-        ], 200
-    else:
-        return generate_racer(req_json["code"], req_json["timeout"])
+    if req_json:
+        if isinstance(req_json["code"], list):
+            return [
+                generate_racer(code, req_json["timeout"])[0]
+                for code in req_json["code"]
+            ], 200
+        else:
+            return generate_racer(req_json["code"], req_json["timeout"])
 
 
 @app.route("/eval_imageoptim_func", methods=["POST"])
 def evaluate_function():
     req_json = request.get_json()
     try:
-        execution_result = sandbox_unsafe_execute(
+        execution_result = unsafe_execute(
             req_json["code"], req_json["func_name"], req_json["timeout"]
         )
         if isinstance(execution_result, ndarray):
@@ -69,14 +69,14 @@ def evaluate_function():
             }, 200
         elif isinstance(execution_result, int):
             return bad_request(
-                f"failed sandbox_unsafe_execute",
+                "Failed sandbox_unsafe_execute",
                 unsafe_execute_error_code=execution_result,
             )
         else:
             bad_request(
-                f"not image", unsafe_execute_error_code=1
+                "Not image", unsafe_execute_error_code=1
             )  # 1: Code runs but fails a test (valid image).
-    except:
+    except Exception:
         return bad_request(
-            f"failed to execute code", unsafe_execute_error_code=6
+            "Failed to execute code", unsafe_execute_error_code=6
         )  # 6: Other errors.
